refactor: utils/helpers + introduce isValidSptId helper
